fix(ci): split oversized test modules
